refactor(core): Split output implementation in smaller modules
