+++ alumet/src/pipeline/elements/error.rs
@@ -19,21 +19,6 @@ pub enum PollError {
     NormalStop,
 }
 
-/// Error which can occur during [`Output::write`](super::super::Output::write).
-#[derive(Debug)]
-pub enum WriteError {
-    /// The measurements could not be written properly, and the output cannot be used anymore.
-    ///
-    /// For instance, a panic may have been caught, or internal data structures may have been messed up.
-    Fatal(anyhow::Error),
-    /// The error is temporary, writing again may work.
-    ///
-    /// You should use this kind of error when:
-    /// - The output communicates with an external entity that you know can fail from time to time.
-    /// - And the output's `write` method can be called again and work. Pay attention to the internal state of the output.
-    CanRetry(anyhow::Error),
-}
-
 impl fmt::Display for PollError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
@@ -44,23 +29,8 @@ impl fmt::Display for PollError {
     }
 }
 
-impl fmt::Display for WriteError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            WriteError::Fatal(e) => write!(f, "fatal error in Output::write: {e}"),
-            WriteError::CanRetry(e) => write!(f, "writing failed (but could work later): {e}"),
-        }
-    }
-}
-
 // Allow to convert from anyhow::Error to pipeline errors
 impl<T: Into<anyhow::Error>> From<T> for PollError {
-    fn from(value: T) -> Self {
-        Self::Fatal(value.into())
-    }
-}
-
-impl<T: Into<anyhow::Error>> From<T> for WriteError {
     fn from(value: T) -> Self {
         Self::Fatal(value.into())
     }
@@ -74,16 +44,5 @@ impl<T, E: Into<anyhow::Error>> PollRetry<T> for Result<T, E> {
     /// Turns this error into [`PollError::CanRetry`].
     fn retry_poll(self) -> Result<T, PollError> {
         self.map_err(|e| PollError::CanRetry(e.into()))
-    }
-}
-
-/// Adds the convenient method `error.retry_write()`.
-pub trait WriteRetry<T> {
-    fn retry_write(self) -> Result<T, WriteError>;
-}
-impl<T, E: Into<anyhow::Error>> WriteRetry<T> for Result<T, E> {
-    /// Turns this error into [`WriteError::CanRetry`].
-    fn retry_write(self) -> Result<T, WriteError> {
-        self.map_err(|e| WriteError::CanRetry(e.into()))
     }
 }
